+++ sqry-core/CHANGELOG.md
@@ -6,7 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-## [12.1.0](https://github.com/verivus-oss/sqry/compare/v12.0.3...v12.1.0) - 2026-05-03
+## [12.1.6](https://github.com/verivus-oss/sqry/compare/v12.1.2...v12.1.6) - 2026-05-04
+
+### Fixed
+- *(index)* stabilize large high-cost graph rebuilds## [12.1.0](https://github.com/verivus-oss/sqry/compare/v12.0.3...v12.1.0) - 2026-05-03
 
 ### Added
 - cross-language field and generic type-parameter emission ([#169](https://github.com/verivus-oss/sqry/pull/169))